Reorganize source folders
