+++ src/Core/Animation/VisualEngine.cs
@@ -6,8 +6,9 @@
 using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using HackenSlay.Core.Objects;
 
-namespace HackenSlay;
+namespace HackenSlay.Core.Animation;
 
 public class VisualEngine
 {
