+++ src/Items/Item.cs
@@ -1,5 +1,6 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using HackenSlay.Core.Objects;
 
 namespace HackenSlay
 {
