+++ src/Items/Weapons/RangedWeapon.cs
@@ -1,9 +1,10 @@
-namespace HackenSlay;
-
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using System;
 using System.Collections.Generic;
+using HackenSlay.Core.Objects;
+
+namespace HackenSlay;
 
 public class RangedWeapon : Weapon
 {
