+++ src/Items/Weapons/Weapon.cs
@@ -1,9 +1,10 @@
-namespace HackenSlay;
-
 using System.Collections;
 using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using HackenSlay.Core.Objects;
+
+namespace HackenSlay;
 
 public class Weapon : TextureObject
 {
