+++ src/Player/Customization/CharacterAppearance.cs
