+++ src/Player/Inventory/Inventory.cs
