+++ src/Player/Player.cs
@@ -11,6 +11,8 @@ using System.Text.Json;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
+using HackenSlay.Core.Objects;
+using HackenSlay.Core.Dev;
 
 namespace HackenSlay;
