+++ src/Tests/HackenSlay.Tests/ItemActionHandlerTests.cs
@@ -2,6 +2,8 @@ using HackenSlay;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Xunit;
+using HackenSlay.Core.Objects;
+using HackenSlay;
 
 namespace HackenSlay.Tests;
